quantumcue-droplet-2-api-server-current/worker/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.3
numpy>=1.26.0
scikit-learn>=1.4.0
Pillow>=10.2.0
python-multipart>=0.0.6
